third_party/WebKit/LayoutTests/fast/dom/DeviceLight/window-property-expected.txt

Tests that the window.DeviceLightEvent and window.ondevicelight properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'DeviceLightEvent' in window is true
PASS window.hasOwnProperty('DeviceLightEvent') is true
PASS typeof window.ondevicelight == 'object' is true
PASS 'ondevicelight' in window is true
PASS successfullyParsed is true

TEST COMPLETE
